requirements.txt
# Core deep learning — ROCm wheels are installed in the rocm/pytorch Docker base image.
# For local (non-Docker) installs, install torch separately:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm6.2
torch>=2.2.0
torchvision>=0.17.0

# Detection model
ultralytics>=8.2.0

# Dataset + augmentation
albumentations>=1.4.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Config + dataset download
PyYAML>=6.0
kagglehub>=0.3.0

# Training QoL
tqdm>=4.65.0
matplotlib>=3.7.0

# Tests
pytest>=7.4.0
